requirements.txt
aiomysql
alembic
anyio
Authlib
bcrypt
beautifulsoup4
certifi
cffi
click
colorama
cryptography
dynaconf
exceptiongroup
fastapi
fastapi-authlib
fastapi-sa
greenlet
gunicorn
h11
httpcore
httpx
idna
inflection
itsdangerous
Jinja2
loguru
Mako
MarkupSafe
motor
pandas
passlib
Pillow
pycparser
pydantic
PyJWT
PyMySQL
python-dotenv
python-jose
python-multipart
requests
selenium
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
websockets
azure-storage-blob